Follow-up to bugfix #9046: Modified type of new field sys_domain.forced
git-svn-id: https://svn.typo3.org/TYPO3v4/Core/trunk@6360 709f56b5-9817-0410-a4d7-c38de5d9e867

+++ typo3/sysext/cms/ext_tables.sql
@@ -310,7 +310,7 @@ CREATE TABLE sys_domain (
   redirectHttpStatusCode int(4) unsigned DEFAULT '301' NOT NULL,
   sorting int(10) unsigned DEFAULT '0' NOT NULL,
   prepend_params int(10) DEFAULT '0' NOT NULL,
-  forced int(10) DEFAULT '0' NOT NULL,
+  forced tinyint(3) unsigned DEFAULT '0' NOT NULL,
 
   PRIMARY KEY (uid),
   KEY parent (pid)
